Damage characters with a shotgun (detailed msg)
I have had to add a collider for the shotgun cone seperately to the cone
visual model so that a script attatched to it can keep track of objects
it's colliding with such that whenever it's called upon (when the player
shoots) it can tell a calling function what it's colliding with. if we
re-used the visual cone, activating the visuals won't provoke a call to
OnTriggerEnter, meaning there's no way to detect collisions

I've had to implement a seperate hitbox on each damagable character
(this is likely for the best anyway) so that the shotgun cone trigger
won't collide with each sub-game object in a character, provoking 3
seperate calls to apply damage on the same parent holding a FPSHealth
object

I've had to give the cone collider a rigidbody so that it isn't counted
as a component of a concave collider for the parent player class, these
colliders look for rigidbodies to make sense of what the collision
means, so I've had to give the cone collider one. If I don't do this,
players can pick up guns by shooting them with a shotgun